refactor: remove 2 tabs
