+++ src/components/Tabs/index.tsx
@@ -14,7 +14,7 @@ type TabsProps = {
 };
 
 const Tabs = ({ tabs }: TabsProps) => {
-  const [activeTab, setActiveTab] = useState(3);
+  const [activeTab, setActiveTab] = useState(2);
   const [hoveredTab, setHoveredTab] = useState<number | null>(null);
 
   return (
