--- src/containers/tabsContainer/Layout/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Layout = () => {
-  return <div>content for Layout</div>;
-};
-
-export default Layout;
