--- src/containers/tabsContainer/Social/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Social = () => {
-  return <div>content for social</div>;
-};
-
-export default Social;
